Activate 'all_platforms' by default.
Had started to allow building only for the local platform
by moving all the platforms that I normally use into a new profile "all_platforms".

Problem:
org.eclipse.core.filesystem.win32.x86 is only used for the Win32/x86 platform.
The org.cs-studio.sns.eclipse.feature also lists it only for that platform, but for some reason maven fails to build when not on win32/x86 because it wants to see that dependency, which wasn't fetched since it's not used..

-> Building for all platforms will result in fetching that depdency, and then it all builds, only actually using org.eclipse.core.filesystem.*.* on the appropriate targets.